third_party/blink/web_tests/fast/forms/select/select-set-length-optgroup-expected.txt

This test that setting HTMLSelectElement.length respects optgroups.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS sel.length is 1
PASS sel.length is 2
PASS og.childElementCount is 1
PASS sel.length is 1
PASS og.childElementCount is 1
PASS sel.length is 1
PASS og.childElementCount is 0
PASS successfullyParsed is true

TEST COMPLETE
